+++ src/main/java/net/daporkchop/fp2/mode/voxel/VoxelPos.java
@@ -26,7 +26,6 @@ import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 import net.daporkchop.fp2.mode.api.IFarPos;
-import net.daporkchop.fp2.mode.heightmap.HeightmapPos;
 import net.minecraft.util.math.AxisAlignedBB;
 
 import java.util.function.Function;
@@ -180,7 +179,7 @@ public class VoxelPos implements IFarPos {
     }
 
     @Override
-    public int localHash() {
+    public long localHash() {
         return interleaveBits(this.x, this.y, this.z);
     }
